Separate the block editor shortcuts from the post editor shortcuts
